feat: add role-based access control with roles and guards for users and notes

+++ backend/src/users/users.service.ts
@@ -2,12 +2,14 @@ import {
   Injectable,
   NotFoundException,
   BadRequestException,
+  ForbiddenException,
 } from '@nestjs/common';
 import { PrismaService } from '../prisma/prisma.service';
 import { UserRequestDto } from './dto/user-request.dto';
 import { UserResponseDto } from './dto/user-response.dto';
 import { User } from '@prisma/client';
 import * as argon from 'argon2';
+import { Role } from '@prisma/client';
 
 @Injectable()
 export class UsersService {
@@ -49,13 +51,24 @@ export class UsersService {
     return this.mapUserToResponse(user);
   }
 
-  async updateUser(id: number, data: UserRequestDto): Promise<UserResponseDto> {
+  async updateUser(
+    id: number,
+    data: UserRequestDto,
+    currentUserId: number,
+  ): Promise<UserResponseDto> {
     // verificar si el usuario existe
     const user = await this.prisma.user.findUnique({
       where: { id },
     });
     if (!user)
       throw new NotFoundException(`El usuario con id ${id} no fue encontrado`);
+
+    // verificar que el usuario no sea administrador
+    if (user.role === Role.ADMIN && id !== currentUserId) {
+      throw new ForbiddenException(
+        'Los administradores no pueden modificarse entre sí',
+      );
+    }
 
     // validar si el email ya está en uso
     const existingUser = await this.prisma.user.findUnique({
@@ -77,13 +90,20 @@ export class UsersService {
     return this.mapUserToResponse(updatedUser);
   }
 
-  async deleteUser(id: number): Promise<void> {
+  async deleteUser(id: number, currentUserId: number): Promise<void> {
     // verificar si el usuario existe
     const user = await this.prisma.user.findUnique({
       where: { id },
     });
     if (!user)
       throw new NotFoundException(`El usuario con id ${id} no fue encontrado`);
+
+    // verificar que el usuario no sea administrador
+    if (user.role === Role.ADMIN && id !== currentUserId) {
+      throw new ForbiddenException(
+        'Los administradores no pueden modificarse entre sí',
+      );
+    }
 
     // eliminar el usuario
     await this.prisma.user.delete({
@@ -101,8 +121,9 @@ export class UsersService {
   mapUserToResponse(user: User): UserResponseDto {
     return {
       id: user.id,
-      email: user.email,
       name: user.name,
+      email: user.email,
+      role: user.role,
       createdAt: user.createdAt,
       updatedAt: user.updatedAt,
     };
